Modified Menu.py
The Menu.py is not a class but a container of functions, there are not many attributes in it. Only things we can change is font (size, color), width, height of screen, maybe add in a background image. The menu is a still page that show buttons to click on and so i figured a class may not be needed. But if you guys think we should make a class to better organize, let me know and i will do it.
For now it works as show when running main.py.